avances cotizar por tipo de servicio

+++ MandaditosExpress/Models/CostoGestionBancaria.cs
@@ -33,11 +33,11 @@ namespace MandaditosExpress.Models
 
         [Required]
         [Display(Name ="Monto Desde")]
-        public decimal MontoDesde { get; set; }
+        public float MontoDesde { get; set; }
 
         [Required]
         [Display(Name = "Monto Hasta")]
-        public decimal MontoHasta { get; set; }
+        public float MontoHasta { get; set; }
 
         [Required]
         public bool Estado { get; set; }
